requirements.txt
# PyTorch and torchvision
torch>=2.0.0
torchvision>=0.15.0

# Image processing and augmentation
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# COCO API for evaluation
pycocotools>=2.0.6

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bar
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Scientific computing
scikit-learn>=1.3.0
scipy>=1.11.0
